pyproject.toml
[project]
name = 'idol_live_website'
version = '0.0.0'
requires-python = '>=3.13'
dependencies = [
  'django==5.2a1',
  'psycopg[binary]==3.2.4',
  'python-dotenv==1.0.1',
  'requests==2.32.3'
]

[tool.distutils.egg_info]
egg_base = "target"